requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.5
django-binary-database-files==1.0.17
filelock==3.8.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.4
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.16.7
Werkzeug==2.2.2
djangorestframework==3.14.0
pytz==2022.7.1
django-filter==22.1
djangorestframework-simplejwt==5.2.2
django-redshift-backend==3.0.0
django-cors-headers==3.13.0
poetry==1.3.2
Pygments==2.14.0
psycopg2==2.9.5
